pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anakin-adk"
version = "0.1.0"
description = "Google ADK tools for web scraping, search, and research powered by Anakin"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Anakin", email = "hello@useanakin.com" }]
keywords = ["google-adk", "adk", "anakin", "web-scraping", "search", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "google-adk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/Anakin-Inc/anakin-adk"
Documentation = "https://docs.useanakin.com"
Repository = "https://github.com/Anakin-Inc/anakin-adk"
Issues = "https://github.com/Anakin-Inc/anakin-adk/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
